Fix fitness precision

+++ Cosette.Tuner.Web/ViewModels/ChromosomeViewModel.cs
@@ -11,7 +11,7 @@ namespace Cosette.Tuner.Web.ViewModels
 
         public DateTime CreationTimeUtc { get; set; }
         public double ElapsedTime { get; set; }
-        public int Fitness { get; set; }
+        public double Fitness { get; set; }
 
         public SelfPlayStatisticsViewModel ReferenceEngineStatistics { get; set; }
         public SelfPlayStatisticsViewModel ExperimentalEngineStatistics { get; set; }
